+++ packages/fabrik_theme/lib/src/responsive/responsive.dart
@@ -1,2 +1,2 @@
-export 'fabrik_breakpoint.dart';
+export 'fabrik_layout.dart';
 export 'fabrik_responsive.dart';